Update src/DkcwdZf2Munee/Controllers/MuneeController.php
Looks like there was an error with a prior rollback where the 'lessifyAllCss' had been set back to true.  This should be false, now fixed.  Additionally inserted exit; to ensure caching is handled properly and that no attempt to render additional content is made.  Thanks to aaron4m for identifying the issue and proposing the change. :-) 